refactor: Refactor code to use new context provider
