+++ src/components/Display/HistoryScreen.js
@@ -1,23 +1,20 @@
-import React from 'react'
-import PropTypes from 'prop-types'
+import React, { useContext } from 'react'
 import { Flex } from 'rebass'
 
+import { CalculatorContext } from '../CalculatorProvider'
 import Row from './Row'
 
-const HistoryScreen = props => (
-  <Flex flexWrap="wrap">
-    {props.history.map((eq, id) => (
-      <Row key={id} id={id}>
-        {eq.join(' ')}
-      </Row>
-    ))}
-  </Flex>
-)
-
-HistoryScreen.propTypes = {
-  history: PropTypes.arrayOf(
-    PropTypes.arrayOf(PropTypes.oneOfType([PropTypes.number, PropTypes.string]))
-  ),
+const HistoryScreen = () => {
+  const { state } = useContext(CalculatorContext)
+  return (
+    <Flex flexWrap="wrap">
+      {state.expHistory.map((eq, id) => (
+        <Row key={id} id={id}>
+          {eq.join(' ')}
+        </Row>
+      ))}
+    </Flex>
+  )
 }
 
 export default HistoryScreen
