+++ src/components/Display/index.js
@@ -1,26 +1,30 @@
-import React from 'react'
+import React, { useContext } from 'react'
 import PropTypes from 'prop-types'
 import { Box, Card } from 'rebass'
 
+import { CalculatorContext } from '../CalculatorProvider'
 import InputScreen from './InputScreen'
 import EquationScreen from './EquationScreen'
 import HistoryScreen from './HistoryScreen'
 
-const Display = props => (
-  <Box p={0}>
-    <Card>
-      <HistoryScreen history={props.history} />
-    </Card>
-    <Card
-      sx={{
-        my: 2,
-      }}
-    >
-      <EquationScreen>{props.equation}</EquationScreen>
-      <InputScreen>{props.input}</InputScreen>
-    </Card>
-  </Box>
-)
+const Display = () => {
+  const { state, display } = useContext(CalculatorContext)
+  return (
+    <Box p={0}>
+      <Card>
+        <HistoryScreen />
+      </Card>
+      <Card
+        sx={{
+          my: 2,
+        }}
+      >
+        <EquationScreen>{state.expression}</EquationScreen>
+        <InputScreen>{display}</InputScreen>
+      </Card>
+    </Box>
+  )
+}
 
 Display.propTypes = {
   history: PropTypes.arrayOf(
