+++ src/components/KeyPad/index.js
@@ -1,5 +1,4 @@
 import React from 'react'
-// import PropTypes from 'prop-types'
 import { Flex, Card } from 'rebass'
 
 import Key from './Key'
