+++ src/pages/index.js
@@ -3,6 +3,7 @@ import { Box } from 'rebass'
 
 import '../fonts/fonts.css'
 import Calculator from '../components/Calculator'
+import CalculatorProvider from '../components/CalculatorProvider'
 
 const Container = props => (
   <Box
@@ -16,11 +17,13 @@ const Container = props => (
 )
 
 const main = () => (
-  <Container>
-    <Box mx="auto" width={256}>
-      <Calculator />
-    </Box>
-  </Container>
+  <CalculatorProvider>
+    <Container>
+      <Box mx="auto" width={256}>
+        <Calculator />
+      </Box>
+    </Container>
+  </CalculatorProvider>
 )
 
 export default main
